housing & degrees privacy info
